+++ BackEnd/src/main/java/br/com/projeto2/aajjl/model/PasswordResetToken.java
@@ -23,6 +23,4 @@ public class PasswordResetToken {
     @OneToOne
     private User user;
 
-    private LocalDateTime expiryDate;
-
 }
