+++ BackEnd/src/main/java/br/com/projeto2/aajjl/service/UserService.java
@@ -2,12 +2,10 @@ package br.com.projeto2.aajjl.service;
 
 
 import br.com.projeto2.aajjl.dto.responses.ResponseDTO;
-import br.com.projeto2.aajjl.model.PasswordResetToken;
 
-import br.com.projeto2.aajjl.model.Patient;
 import br.com.projeto2.aajjl.model.Profession;
 import br.com.projeto2.aajjl.model.User;
-import br.com.projeto2.aajjl.repository.PasswordResetTokenRepository;
+
 import br.com.projeto2.aajjl.repository.UserRepository;
 import br.com.projeto2.aajjl.security.TokenService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -15,10 +13,9 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.lang.reflect.Field;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
-import java.util.UUID;
+
 
 @Service
 public class UserService {
@@ -34,9 +31,6 @@ public class UserService {
 
     @Autowired
     private TokenService tokenService;
-
-    @Autowired
-    private PasswordResetTokenRepository tokenRepository;
 
     public ResponseDTO create(User newUser) {
 
@@ -68,7 +62,6 @@ public class UserService {
         );
 
     }
-
 
     public List<User> getAll() {
         return userRepository.findAll()
@@ -148,49 +141,5 @@ public class UserService {
     public List<User> getAllInativos() {
         return userRepository.findByAtivoFalse();
     }
-
-
-    public void sendPasswordResetToken(String email) {
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
-
-        //Gera o token
-        String token = UUID.randomUUID().toString();
-
-        PasswordResetToken resetToken = new PasswordResetToken();
-        resetToken.setToken(token);
-        resetToken.setUser(user);
-        resetToken.setExpiryDate(LocalDateTime.now().plusMinutes(30));
-
-        tokenRepository.save(resetToken);
-
-        //Conferir link se é esse mesmo, e lembrar de trocar no host
-        String link = "https://localhost:8080/reset-password?token=" + token;
-
-        String assunto = "Recuperação de Senha";
-        String mensagem = "Olá, " + user.getNome() + ".\n\n" +
-                "Clique no link abaixo para redefinir sua senha:\n" +
-                link + "\n\n" +
-                "Este link expirará em 30 minutos.";
-
-        emailService.enviarEmailSimples(user.getEmail(), assunto, mensagem);
-    }
-
-    public void resetPassword(String token, String newPassword) {
-        PasswordResetToken resetToken = tokenRepository.findByToken(token)
-                .orElseThrow(() -> new RuntimeException("Token inválido"));
-
-        if (resetToken.getExpiryDate().isBefore(LocalDateTime.now())) {
-            throw new RuntimeException("Token expirado");
-        }
-
-        User user = resetToken.getUser();
-        user.setSenha(newPassword);
-        userRepository.save(user);
-
-        tokenRepository.delete(resetToken);
-    }
-
-
 
 }
